Rename to encoding.py
