+++ fauna/encoding.py
